Land "Adding multigammaln ref"
Partially based on https://github.com/pytorch/pytorch/pull/83662.

I'll help land this one, as Rob does not work in the PyTorch project
anymore

I removed the data-dependent check for the args, as data dependencies
are bad for many reasons (and it was failing when the input has NaNs).

It also registers arange as a decomposition, and fixes the naming of its
args.

ghstack-source-id: 6d842325a91f00b2af3ee15bf8284c4d1be7e868
Pull Request resolved: https://github.com/pytorch/pytorch/pull/85153